mv: don't rename files onto the place of directories (should create files in directories eventually)

+++ sys/mv.sl
@@ -1,5 +1,6 @@
 include "stdio.sl";
 include "sys.sl";
+include "malloc.sl";
 
 var args = cmdargs()+1;
 
@@ -10,7 +11,16 @@ if (!*args || !*(args+1) || *(args+2)) {
 
 # TODO: [nice] if *(args+1) is a directory, create the name inside the directory
 
-var n = rename(*args, *(args+1));
+var n;
+
+var statbuf = malloc(4);
+n = stat(*(args+1), statbuf);
+if (n == 0 && statbuf[0] == 0) {
+    fprintf(2, "mv: %s: is directory\n", [*(args+1)]);
+    exit(1);
+};
+
+n = rename(*args, *(args+1));
 if (n == EXISTS) {
     unlink(*(args+1));
     n = rename(*args, *(args+1));
